Doc: refine logging conventions
There were some ambiguities in the logging conventions, that made it difficult to distinguish when to use log levels 2, 3 and 4 respectively. 

This PR refines the descriptions of each log level so that it is clearer when each level should be applied.

An additional rule is added for logging before actions in the present continuous tense.