fix: deletion was blocked when other clusters are not ready
An internal annotation 'pending-sync-clusters' is introduced to
record which member clusters the sync controller needs to deliver
resources to. When the dispatch is successful, the record of
the member cluster will be removed in 'pending-sync-clusters'.
So unrelated clusters that are not ready will not prevent the deletion
of federated resources.